requirements.txt
toml
pyspark
pandas
numpy
ipykernel
Faker
mysql-connector-python
python-dotenv
PyYAML
tqdm
requests
apache-airflow
-e .